Revert "Merge remote-tracking branch 'upstream/develop' into BenediktSailer"
This reverts commit c80f88f3c21fb7f1a0102e00225575c4a2d62502, reversing
changes made to c0d82e6fd3d0874209f682765733a346bc0336c9.

+++ modules/juce_audio_processors/format_types/VST3_SDK/public.sdk/source/vst/vstpresetfile.cpp
@@ -8,7 +8,7 @@
 //
 //-----------------------------------------------------------------------------
 // LICENSE
-// (c) 2021, Steinberg Media Technologies GmbH, All Rights Reserved
+// (c) 2019, Steinberg Media Technologies GmbH, All Rights Reserved
 //-----------------------------------------------------------------------------
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -712,7 +712,7 @@ IMPLEMENT_FUNKNOWN_METHODS (FileStream, IBStream, IBStream::iid)
 //------------------------------------------------------------------------
 tresult PLUGIN_API FileStream::read (void* buffer, int32 numBytes, int32* numBytesRead)
 {
-	size_t result = fread (buffer, 1, static_cast<size_t> (numBytes), file);
+	size_t result = fread (buffer, 1, numBytes, file);
 	if (numBytesRead)
 		*numBytesRead = (int32)result;
 	return static_cast<int32> (result) == numBytes ? kResultOk : kResultFalse;
@@ -721,7 +721,7 @@ tresult PLUGIN_API FileStream::read (void* buffer, int32 numBytes, int32* numByt
 //------------------------------------------------------------------------
 tresult PLUGIN_API FileStream::write (void* buffer, int32 numBytes, int32* numBytesWritten)
 {
-	size_t result = fwrite (buffer, 1, static_cast<size_t> (numBytes), file);
+	size_t result = fwrite (buffer, 1, numBytes, file);
 	if (numBytesWritten)
 		*numBytesWritten = (int32)result;
 	return static_cast<int32> (result) == numBytes ? kResultOk : kResultFalse;
@@ -863,9 +863,9 @@ BufferStream::~BufferStream () {FUNKNOWN_DTOR}
 //------------------------------------------------------------------------
 tresult PLUGIN_API BufferStream::read (void* buffer, int32 numBytes, int32* numBytesRead)
 {
-	uint32 size = mBuffer.get (buffer, static_cast<uint32> (numBytes));
+	uint32 size = mBuffer.get (buffer, numBytes);
 	if (numBytesRead)
-		*numBytesRead = static_cast<int32> (size);
+		*numBytesRead = size;
 
 	return kResultTrue;
 }
@@ -873,7 +873,7 @@ tresult PLUGIN_API BufferStream::read (void* buffer, int32 numBytes, int32* numB
 //------------------------------------------------------------------------
 tresult PLUGIN_API BufferStream::write (void* buffer, int32 numBytes, int32* numBytesWritten)
 {
-	bool res = mBuffer.put (buffer, static_cast<uint32> (numBytes));
+	bool res = mBuffer.put (buffer, numBytes);
 	if (numBytesWritten)
 		*numBytesWritten = res ? numBytes : 0;
